Rewrite sentence building app (mobile-friendly) ++
Previously, to select word you should've drag cursor. Now you need
to click the start and the end positions.
This was made in preparation to later addition of 'manual word choosing'
Now we can only add words automatically via mecab; because of that
some words cannot be added (eg.: if word written in kana [かしら／頭],
or when we needed more control)